Edit user profile
In this implementation, the following features are included:

Flash Message on Registration:
During the user registration process, if a duplicate username or email is encountered, a flash message is displayed. This informs the user to choose a different username or email.
Edit User Profile:
Users can successfully edit all their profile information. This includes changing their email and password, after which they are required to log in again to validate the changes.
Edit User Profile
Username Duplication:
When a user tries to change their username, if the new username already exists, the change is not allowed. However, no error flash message is displayed in this case.
Edit User Profile
Change Password:
When a user attempts to change their password, if the new password confirmation does not match the previous one, the update is not allowed. Again, no error flash message is displayed in this scenario.